[major] adds a new data fingerprinting module
Can be used to scan for software or data fingerprints on the filesystem,
returning information related to matches against the fingerprint. The
initial example included identifies the version of MediaWiki deployments
that exist on the device MIG is running on, but this can be expanded to
include any other file content type fingerprint.